Added a factory method that supports BASIC auth.
My original requirement is to be able to send an HTTP basic
authentication along with requests.

I took the liberty to also add a version of a method that takes
arbitrary RequestInterceptors to support other possible modes of
authentication, and general control. I think it's worthwhile doing this
now to avoid ending up with 20 different overloaded flavors of the
getInstance() method, but I can imagine you might not want to expose
Feign or perhaps go the other way and not have the
getInstanceWithBasicAuth method.