Allow updating of an existing application group.
Marathon allows application groups to be updated, possibly inducing a
new deployment to start. The PUT resource is documented here:

  https://mesosphere.github.io/marathon/docs/rest-api.html#put-v2-groups-groupid

Note that even though the documentation doesn't mention it, this
resource accepts a "dry run" flag, per the definition here:

  https://github.com/mesosphere/marathon/blob/master/src/main/scala/mesosphere/marathon/api/v2/GroupsResource.scala#L118

Hence the third parameter on the new Marathon#updateGroup() method.